Scan refactor and many fixes
