+++ TVRename/Forms/ShowPreferences/QuickLocate.cs
@@ -1,11 +1,12 @@
+// 
+// Main website for TVRename is http://tvrename.com
+// 
+// Source code available at https://github.com/TV-Rename/tvrename
+// 
+// This code is released under GPLv3 https://github.com/TV-Rename/tvrename/blob/master/LICENSE.md
+// 
+
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using Alphaleonis.Win32.Filesystem;
 
@@ -25,7 +26,7 @@ namespace TVRename.Forms.ShowPreferences
             cbDirectory.SelectedIndex = 0;
             cbDirectory.ResumeLayout();
 
-            txtShowFolder.Text = Path.DirectorySeparatorChar + TVSettings.Instance.FilenameFriendly(FileHelper.MakeValidPath(hint)) + Path.DirectorySeparatorChar;
+            txtShowFolder.Text = Path.DirectorySeparatorChar + TVSettings.Instance.FilenameFriendly(FileHelper.MakeValidPath(hint));
         }
 
         private void buttonOK_Click(object sender, EventArgs e)
