+++ TVRename/Forms/UI.cs
@@ -932,7 +932,7 @@ namespace TVRename
                 return;
 
             ProcessedEpisode ei = (ProcessedEpisode) lvWhenToWatch.SelectedItems[0].Tag;
-            List<FileInfo> fl = TVDoc.FindEpOnDisk(null, ei);
+            List<FileInfo> fl = FinderHelper.FindEpOnDisk(null, ei);
             if (fl != null && fl.Count > 0)
             {
                 Helpers.SysOpen(fl[0].FullName);
@@ -1246,7 +1246,7 @@ namespace TVRename
 
             if (ep != null && mLastShowsClicked != null && mLastShowsClicked.Count == 1)
             {
-                List<FileInfo> fl = TVDoc.FindEpOnDisk(null, ep);
+                List<FileInfo> fl = FinderHelper.FindEpOnDisk(null, ep);
                 if (fl == null) return;
 
                 if (fl.Count <= 0) return;
@@ -1271,7 +1271,7 @@ namespace TVRename
                 bool first = true;
                 foreach (ProcessedEpisode epds in si.SeasonEpisodes[seas.SeasonNumber])
                 {
-                    List<FileInfo> fl = TVDoc.FindEpOnDisk(null, epds);
+                    List<FileInfo> fl = FinderHelper.FindEpOnDisk(null, epds);
                     if (fl != null && fl.Count > 0)
                     {
                         if (first)
@@ -2019,7 +2019,7 @@ namespace TVRename
 
             if (airdt.Value.CompareTo(DateTime.Now) < 0) // has aired
             {
-                List<FileInfo> fl = TVDoc.FindEpOnDisk(dfc, pe);
+                List<FileInfo> fl = dfc.FindEpOnDisk(pe);
                 if (fl != null && fl.Count > 0)
                     lvi.ImageIndex = 0;
                 else if (pe.Show.DoMissingCheck)
@@ -2497,14 +2497,14 @@ namespace TVRename
                 Logger.Info("Parsing {0} for new shows", dirPath);
                 if (!Directory.Exists(dirPath)) continue;
                 try
-                { 
+                {
                     foreach (string filePath in Directory.GetFiles(dirPath, "*", SearchOption.AllDirectories))
                     {
                         if (!File.Exists(filePath)) continue;
 
                         FileInfo fi = new FileInfo(filePath);
 
-                        if (FileHelper.IgnoreFile(fi)) continue;
+                        if (fi.IgnoreFile()) continue;
 
                         if (!LookForSeries(fi.Name)) possibleShowNames.Add(fi.RemoveExtension() + ".");
                     }
@@ -2542,10 +2542,10 @@ namespace TVRename
                 }
 
                 //Remove any (nnnn) in the hint - probably a year
-                string refinedHint = Regex.Replace(hint,@"\(\d{4}\)","");
+                string refinedHint = Regex.Replace(hint, @"\(\d{4}\)", "");
 
                 //Remove anything we can from hint to make it cleaner and hence more likely to match
-                refinedHint = RemoveSeriesEpisodeIndicators(refinedHint);
+                refinedHint = Helpers.RemoveSeriesEpisodeIndicators(refinedHint, mDoc.Library.SeasonWords());
 
                 if (string.IsNullOrWhiteSpace(refinedHint))
                 {
@@ -2606,63 +2606,6 @@ namespace TVRename
         {
             return TVSettings.Instance.AutoAddMovieTermsArray.Any(term =>
                 hint.Contains(term, StringComparison.OrdinalIgnoreCase));
-        }
-
-        private string RemoveSeriesEpisodeIndicators(string hint)
-        {
-            string hint2 = Helpers.RemoveDiacritics(hint);
-            hint2 = RemoveSe(hint2);
-            hint2 = hint2.ToLower();
-            hint2 = hint2.Replace("'", "");
-            hint2 = hint2.Replace("&", "and");
-            hint2 = Regex.Replace(hint2, "[_\\W]+", " ");
-            foreach (string term in TVSettings.Instance.AutoAddIgnoreSuffixesArray)
-            {
-                hint2 = hint2.RemoveAfter(term);
-            }
-
-            foreach (string seasonWord in mDoc.Library.SeasonWords())
-            {
-                hint2 = hint2.RemoveAfter(seasonWord);
-            }
-
-            hint2 = hint2.Trim();
-            return hint2;
-        }
-
-        private static string RemoveSe(string hint)
-        {
-            foreach (TVSettings.FilenameProcessorRE re in TVSettings.Instance.FNPRegexs)
-            {
-                if (!re.Enabled)
-                    continue;
-
-                try
-                {
-                    Match m = Regex.Match(hint, re.RegExpression, RegexOptions.IgnoreCase);
-                    if (m.Success)
-                    {
-                        if (!int.TryParse(m.Groups["s"].ToString(), out int seas))
-                            seas = -1;
-
-                        if (!int.TryParse(m.Groups["e"].ToString(), out int ep))
-                            ep = -1;
-
-                        int p = Math.Min(m.Groups["s"].Index, m.Groups["e"].Index);
-                        int p2 = Math.Min(p, hint.IndexOf(m.Groups.SyncRoot.ToString(), StringComparison.Ordinal));
-
-                        if (seas != -1 && ep != -1) return hint.Remove(p2 != -1 ? p2 : p);
-                    }
-                }
-                catch (FormatException)
-                {
-                }
-                catch (ArgumentException)
-                {
-                }
-            }
-
-            return hint;
         }
 
         private bool LookForSeries(string name) => LookForSeries(name, mDoc.Library.Values);
@@ -3419,7 +3362,7 @@ namespace TVRename
 
         private void duplicateFinderLOGToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            List<PossibleDuplicateEpisode> x = mDoc.FindDoubleEps();
+            List<PossibleDuplicateEpisode> x = Beta.FindDoubleEps(mDoc);
             DupEpFinder form = new DupEpFinder(x, mDoc, this);
             form.ShowDialog();
         }
@@ -3511,7 +3454,7 @@ namespace TVRename
 
         private void episodeFileQualitySummaryLogToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            mDoc.LogShowEpisodeSizes();
+            Beta.LogShowEpisodeSizes(mDoc);
         }
     }
 }
