+++ TVRename/ItemsAndActions/ActionTDownload.cs
@@ -87,7 +87,7 @@ namespace TVRename
 
         private static string SaveDownloadedData(byte[] r,string name)
         {
-            string saveTemp = Path.GetTempPath() + System.IO.Path.DirectorySeparatorChar + TVSettings.Instance.FilenameFriendly(name);
+            string saveTemp = Path.GetTempPath() + Path.DirectorySeparatorChar + TVSettings.Instance.FilenameFriendly(name);
             if (new FileInfo(saveTemp).Extension.ToLower() != "torrent")
                 saveTemp += ".torrent";
             File.WriteAllBytes(saveTemp, r);
