+++ TVRename/Settings/Settings.cs
@@ -168,7 +168,7 @@ namespace TVRename
         public BetaMode mode = BetaMode.ProductionOnly;
         public float upgradeDirtyPercent = 20;
         public float replaceMargin = 10;
-        public bool ReplaceWithBetterQuality = true;
+        public bool ReplaceWithBetterQuality = false;
         public KeepTogetherModes keepTogetherMode = KeepTogetherModes.All;
 
         public bool BulkAddIgnoreRecycleBin = false;
@@ -257,6 +257,7 @@ namespace TVRename
         public bool runPeriodicCheck = false;
         public int periodCheckHours = 1;
         public bool runStartupCheck = false;
+        public bool DoBulkAddInScan = false; 
 
         private TVSettings()
         {
@@ -398,6 +399,7 @@ namespace TVRename
             XmlHelper.WriteElementToXml(writer, "PeriodicScan", runPeriodicCheck);
             XmlHelper.WriteElementToXml(writer, "PeriodicScanHours", periodCheckHours);
             XmlHelper.WriteElementToXml(writer, "RemoveDownloadDirectoriesFiles", RemoveDownloadDirectoriesFiles);
+            XmlHelper.WriteElementToXml(writer, "DoBulkAddInScan", DoBulkAddInScan);
             XmlHelper.WriteElementToXml(writer, "DeleteShowFromDisk", DeleteShowFromDisk);
             XmlHelper.WriteElementToXml(writer, "SABAPIKey", SABAPIKey);
             XmlHelper.WriteElementToXml(writer, "CheckSABnzbd", CheckSABnzbd);
@@ -944,7 +946,7 @@ namespace TVRename
             BGDownload = xmlSettings.ExtractBool("BGDownload") ?? false;
             OfflineMode = xmlSettings.ExtractBool("OfflineMode")??false;
             DetailedRSSJSONLogging = xmlSettings.ExtractBool("DetailedRSSJSONLogging") ?? false;
-            ReplaceWithBetterQuality = xmlSettings.ExtractBool("ReplaceWithBetterQuality")??true;
+            ReplaceWithBetterQuality = xmlSettings.ExtractBool("ReplaceWithBetterQuality")??false;
             ExportWTWRSSTo = xmlSettings.ExtractString("ExportWTWRSSTo");
             ExportWTWXML = xmlSettings.ExtractBool("ExportWTWXML")??false;
             ExportWTWXMLTo = xmlSettings.ExtractString("ExportWTWXMLTo");
@@ -1050,6 +1052,7 @@ namespace TVRename
             runPeriodicCheck = xmlSettings.ExtractBool("PeriodicScan") ?? false;
             periodCheckHours = xmlSettings.ExtractInt("PeriodicScanHours")??1;
             RemoveDownloadDirectoriesFiles = xmlSettings.ExtractBool("RemoveDownloadDirectoriesFiles") ?? false;
+            DoBulkAddInScan = xmlSettings.ExtractBool("DoBulkAddInScan") ?? false;
             DeleteShowFromDisk = xmlSettings.ExtractBool("DeleteShowFromDisk")??true;
             EpJPGs = xmlSettings.ExtractBool("EpJPGs") ?? false;
             SeriesJpg = xmlSettings.ExtractBool("SeriesJpg") ?? false;
