+++ TVRename/Utility/SetProgressDelegate.cs
@@ -1,4 +1,12 @@
+// 
+// Main website for TVRename is http://tvrename.com
+// 
+// Source code available at https://github.com/TV-Rename/tvrename
+// 
+// This code is released under GPLv3 https://github.com/TV-Rename/tvrename/blob/master/LICENSE.md
+// 
+
 namespace TVRename
 {
-    public delegate void SetProgressDelegate(int percent);
+    public delegate void SetProgressDelegate(int percent,string message);
 }
